Minimize host-verify package without separating oehost
Make possible to use cpack to create a usable open-enclave-hostverify package with necessary header files and targets;
that is, we can build the host-verify sample without any modification of files installed from open-enclave-hostverify.

This is a workaround based on pull request #3123 and issues #3160 and #3181 so that we do not have to change any logic or testing flow for now.

Signed-off-by: Ryan Hsu <ryhsu@microsoft.com>